signalgenerator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pymongo==4.6.0
python-dotenv==1.0.0

